\name{enirg.GUI}
\alias{enirg.GUI}

\title{
Graphical User Interface of the \emph{ENiRG} package 
}

\description{
\code{enirg.GUI} creates a friendly GUI to set functions available in this package. Interface relies on \emph{tcltk} and \emph{fgui} routines to parse functions and create widgets.
}

\usage{
enirg.GUI()
}

\value{
Displays a graphic user interface.
}

\author{
Fernando Canovas \email{fcgarcia@ualg.pt}, Chiara  Magliozzi \email{chiara.magliozzi@libero.it}
}


\examples{
enirg.GUI()
}

\keyword{enirg}
\keyword{gui}
